Restruct of class
implemnted singlton, cli is working with new set up, need to fix minor checks for conditions. Also need to fix for GUI
